+++ R/class_url.R
@@ -37,9 +37,10 @@ store_coerce_object.tar_url <- function(store, object) {
 }
 
 #' @export
-store_assert_format.tar_url <- function(store, object) {
-  if (!is.null(object) && !is.character(object)) {
+store_assert_format.tar_url <- function(store, object, name) {
+  if (!is.character(object)) {
     throw_validate(
+      "target ", name, " did not return a character. ",
       "targets with format = \"url\" must return ",
       "character vectors of URL paths."
     )